requirements.txt
torch
torchvision
diffusers
rembg
trimesh
pillow
imageio
accelerate
transformers
onnxruntime